requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
attrs==23.2.0
cachetools==5.5.2
certifi==2024.8.30
chardet==5.2.0
charset-normalizer==3.3.2
colorama==0.4.6
Django==4.2.11
django-ckeditor==6.7.1
django-colorfield==0.11.0
django-cors-headers==4.3.1
django-des==2.4.1
django-filter==24.1
django-jazzmin==3.0.0
django-js-asset==2.2.0
django-monthfield==0.1.4
django-multiselectfield==0.1.12
django-phonenumber-field==8.0.0
django-seed==0.3.1
django-solo==1.1.3
django-wkhtmltopdf==3.4.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.1
et-xmlfile==1.1.0
Faker==24.9.0
google-ai-generativelanguage==0.6.15
google-api-core==2.24.2
google-api-python-client==2.167.0
google-auth==2.39.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
grpcio==1.71.0
grpcio-status==1.71.0
httplib2==0.22.0
idna==3.8
inflection==0.5.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
marshmallow==3.22.0
oauthlib==3.2.2
openpyxl==3.1.2
packaging==24.1
pdfkit==1.0.0
phonenumbers==8.13.44
pillow==10.3.0
proto-plus==1.26.1
protobuf==5.29.4
psycopg-binary==3.2.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.5
pydantic_core==2.33.2
PyJWT==2.8.0
PyMuPDF==1.26.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
PyYAML==6.0.1
referencing==0.34.0
reportlab==4.2.0
requests==2.32.3
requests-oauthlib==2.0.0
rpds-py==0.18.0
rsa==4.9.1
six==1.16.0
sqlparse==0.4.4
sslcommerz-lib==1.0
sslcommerz-sdk==1.0.4
toposort==1.10
tqdm==4.67.1
typed-ast==1.5.5
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
